refactor: move all components related to home page
