+++ components/home/guarantees/components/card/card.tsx
