+++ components/home/preparation/preparation.tsx
